refactor: split layout/mod.rs into pane/tree/preset/split/tab sub-modules
- pane.rs: Pane, PaneSource, Selection types (~136 LOC)
- tree.rs: PaneNode, SplitDir, tree transforms, swap (~317 LOC)
- preset.rs: LayoutPreset, build_tree, build_preset (~123 LOC)
- split.rs: ratio_to_size, hit-testing, resize, Direction (~255 LOC)
- tab.rs: DragTabTarget, Tab struct + methods (~337 LOC)
- mod.rs: Layout struct, resize_panes, re-exports (~258 LOC)

All pub types re-exported via mod.rs — zero caller import changes.
